refactor: move shared domain types to models/ package
Extract Allocation, Installment, Settlement, AnticipationResult, and
BillingCycleLoanStatement into money_warp/models/ to break the circular
dependency between engines/ and loan/.

engines/ now imports domain types from models/ instead of loan/,
eliminating the need for the PEP 562 __getattr__ lazy-import hack
in loan/__init__.py. Loan is now eagerly imported.

Backward-compatible: loan/__init__.py and billing_cycle_loan/__init__.py
re-export the types from models/, so existing import paths still work.
