+++ knowledge/architecture.md
@@ -7,6 +7,12 @@ MoneyWarp is a Time Value of Money library built around the metaphor of "time-wa
 ```
 money_warp/
 ├── __init__.py            # Public API exports
+├── models/
+│   ├── __init__.py        # Re-exports all shared domain types
+│   ├── allocation.py      # Allocation dataclass (per-installment payment breakdown)
+│   ├── installment.py     # Installment dataclass
+│   ├── settlement.py      # Settlement, AnticipationResult dataclasses
+│   └── statement.py       # BillingCycleLoanStatement dataclass
 ├── engines/
 │   ├── __init__.py        # Re-exports public API (single import path)
 │   ├── interest.py        # InterestCalculator, MoraStrategy, MoraRateCallback
@@ -28,18 +34,14 @@ money_warp/
 ├── billing_cycle_loan/
 │   ├── billing_cycle_loan.py  # BillingCycleLoan facade
 │   ├── engines.py             # BCL-specific: mora rate resolution + statements
-│   ├── mora_rate_resolver.py  # MoraRateResolver protocol
-│   └── statement.py           # BillingCycleLoanStatement dataclass
+│   └── mora_rate_resolver.py  # MoraRateResolver protocol
 ├── credit_card/
 │   ├── credit_card.py     # CreditCard state machine
 │   └── statement.py       # Statement (frozen derived view)
 ├── loan/
 │   ├── loan.py            # Loan facade
 │   ├── engines.py         # Backward-compatible re-exports from money_warp.engines
-│   ├── tvm.py             # TVM standalone functions (PV, IRR, anticipation)
-│   ├── allocation.py      # Allocation dataclass (per-installment payment breakdown)
-│   ├── installment.py     # Installment dataclass
-│   └── settlement.py      # Settlement, AnticipationResult dataclasses
+│   └── tvm.py             # TVM standalone functions (PV, IRR, anticipation)
 ├── scheduler/
 │   ├── base.py            # BaseScheduler (abstract)
 │   ├── price_scheduler.py           # PriceScheduler (French amortization)
@@ -101,9 +103,10 @@ Equality between `CashFlowItem` and `CashFlowEntry` uses Python's reflected-equa
 
 `Loan` is a facade that orchestrates three focused components:
 
-- **`engines/`** (package) — all shared computation: `InterestCalculator`, `MoraStrategy`, `MoraRateCallback` (interest primitives), `compute_state` (unified forward pass with optional `mora_rate_for_event` callback), `allocate_payment` (fine -> mora -> interest -> principal priority), `compute_fines_at`, `apply_tolerance_adjustment`, `build_installments`. Split into submodules (`interest.py`, `allocation.py`, `fines.py`, `forward_pass.py`) with a single import path via `__init__.py`.
+- **`models/`** (package) — shared domain result types (`Allocation`, `Installment`, `Settlement`, `AnticipationResult`, `BillingCycleLoanStatement`) used by both `engines/` and product packages. These are leaf-level frozen dataclasses with no dependency on `Loan` or engine logic.
+- **`engines/`** (package) — all shared computation: `InterestCalculator`, `MoraStrategy`, `MoraRateCallback` (interest primitives), `compute_state` (unified forward pass with optional `mora_rate_for_event` callback), `allocate_payment` (fine -> mora -> interest -> principal priority), `compute_fines_at`, `apply_tolerance_adjustment`, `build_installments`. Split into submodules (`interest.py`, `allocation.py`, `fines.py`, `forward_pass.py`) with a single import path via `__init__.py`. Imports domain types from `models/`.
 - **`loan/engines.py`** — backward-compatible re-export shim for existing code importing from `money_warp.loan.engines`.
-- **`tvm.py`** — standalone functions for PV, IRR, and anticipation (eliminates circular imports).
+- **`tvm.py`** — standalone functions for PV, IRR, and anticipation.
 
 `Loan.cashflow` is the single source of truth — expected schedule items and actual payment items live in one `CashFlow`. All derived state (settlements, installments, balances, fines) is computed on demand via a forward pass. Schedule generation (`generate_expected_cash_flow`, `get_original_schedule`, `get_amortization_schedule`) stays in `Loan`.
