+++ money_warp/billing_cycle_loan/__init__.py
@@ -1,8 +1,8 @@
 """Billing-cycle loan — fixed amortization with billing-cycle payment timing."""
 
+from ..models import BillingCycleLoanStatement
 from .billing_cycle_loan import BillingCycleLoan
 from .mora_rate_resolver import MoraRateResolver
-from .statement import BillingCycleLoanStatement
 
 __all__ = [
     "BillingCycleLoan",
