+++ money_warp/billing_cycle_loan/engines.py
@@ -13,11 +13,10 @@ from ..billing_cycle import BaseBillingCycle
 from ..engines import InterestCalculator, LoanState, MoraRateCallback
 from ..engines import compute_state as _compute_state
 from ..interest_rate import InterestRate
-from ..loan.settlement import Settlement
+from ..models import BillingCycleLoanStatement, Settlement
 from ..money import Money
 from ..scheduler import PaymentSchedule
 from .mora_rate_resolver import MoraRateResolver
-from .statement import BillingCycleLoanStatement
 
 # ------------------------------------------------------------------
 # Mora rate resolution
