+++ money_warp/engines/forward_pass.py
@@ -12,9 +12,9 @@ from typing import Dict, List, Optional, Tuple
 
 from ..cash_flow import CashFlow, CashFlowItem
 from ..interest_rate import InterestRate
-from ..loan.allocation import Allocation
-from ..loan.installment import Installment
-from ..loan.settlement import Settlement
+from ..models.allocation import Allocation
+from ..models.installment import Installment
+from ..models.settlement import Settlement
 from ..money import Money
 from ..scheduler import PaymentSchedule, PaymentScheduleEntry
 from ..time_context import TimeContext
