+++ money_warp/models/__init__.py
@@ -1,0 +1,14 @@
+"""Shared domain model types used across loan products and engines."""
+
+from .allocation import Allocation
+from .installment import Installment
+from .settlement import AnticipationResult, Settlement
+from .statement import BillingCycleLoanStatement
+
+__all__ = [
+    "Allocation",
+    "AnticipationResult",
+    "BillingCycleLoanStatement",
+    "Installment",
+    "Settlement",
+]
